+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - Make some `NablaTheme` property public instead of internal.
 - A `NetworkError` will now be returned in case of a network error instead of a `ServerError` before.
+- Fix an issue where date separators would be incorrectly ordered in the chat.
 
 ## [1.0.0-alpha09] - 2022-07-13